move buttons components to one directory
