+++ src/common/buttons/icon-button/styles.scss
