+++ src/common/modal-containers/destructive/index.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 
 import { CommonModal } from "common/modal";
-import { ActionButton } from "common/action-button";
+import { ActionButton } from "common/buttons/action-button";
 
 import "./styles.scss";
